% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunaRity-package.R
\docType{package}
\name{lacunaRity-package}
\alias{lacunaRity}
\alias{lacunaRity-package}
\title{lacunaRity: Fast 3D Lacunarity for LiDAR Voxel Data}
\description{
Calculates lacunarity curves for 3D voxel data using a fast C++ implementation of the gliding-box algorithm. Gliding-box lacunarity can be calculated using the traditional fixed-boundary algorithm or the periodic-boundary algorithm introduced by Feagin et al 2006.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElliottSmeds/lacunaRity}
  \item Report bugs at \url{https://github.com/ElliottSmeds/lacunaRity/issues}
}

}
\author{
\strong{Maintainer}: Elliott Smeds \email{elliott.alfred93@gmail.com} (\href{https://orcid.org/0000-0003-1054-7491}{ORCID})

}
\keyword{internal}
