% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunarity.R
\name{lacunarity}
\alias{lacunarity}
\title{Gliding-box lacunarity}
\usage{
lacunarity(x, box_sizes = "twos")
}
\arguments{
\item{x}{a 3-dimensional \code{\link[=array]{array}} of integer values}

\item{box_sizes}{Which box sizes to use for calculating lacunarity:
\itemize{
\item "twos" returns box sizes for all powers of two less
than or equal to the smallest dimension of \code{x}.
\item "all" calculates every possible box size up to the smallest dimension of \code{x}.
}}
}
\value{
a \code{\link[=data.frame]{data.frame}} containing box sizes and their
corresponding lacunarity values
}
\description{
Gliding-box lacunarity
}
\examples{
a <- array(data = rep(c(1,0), 125), dim = c(5,5,5))
lacunarity(a)
}
