% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_box.R
\name{bounding_box}
\alias{bounding_box}
\alias{bounding_box.default}
\alias{bounding_box.lac_voxels}
\title{Create voxel array}
\usage{
bounding_box(x, threshold = 0, edge_length = NULL)
}
\arguments{
\item{x}{text}

\item{threshold}{text}

\item{edge_length}{text}
}
\value{
A 3-dimensional integer \code{\link[=array]{array}} containing values of \code{0}
or \code{1}, representing the occupancy of a given voxel
}
\description{
Converts a table of voxel data into a 3-dimensional array
}
\examples{
# simulate a diagonal line of points with XYZ coordinates
pc <- data.frame(X = 0:99, Y = 0:99, Z = 0:99)
# convert point data to cubic voxels of length 5
vox <- voxelize(pc, edge_length = c(5,5,5))
# convert to voxel array
bounding_box(vox)
}
