\name{lacunaRity-package}
\alias{lacunaRity-package}
\alias{lacunaRity}
\docType{package}
\title{
\packageTitle{lacunaRity}
}
\description{
\packageDescription{lacunaRity}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lacunaRity}
\packageIndices{lacunaRity}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lacunaRity}

Maintainer: \packageMaintainer{lacunaRity}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
