% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunarity.R
\name{lacunarity}
\alias{lacunarity}
\title{Calculate gliding-box lacunarity}
\usage{
lacunarity(x, box_sizes = "twos", periodic = FALSE)
}
\arguments{
\item{x}{A 3-dimensional \code{\link[=array]{array}} of integer values}

\item{box_sizes}{Which box sizes to use for calculating lacunarity:
\itemize{
\item \code{"twos"} (the default) returns box sizes for all powers of two less than or
equal to the smallest dimension of \code{x}.
\item \code{"all"} calculates every possible box size up to the smallest dimension of \code{x}.
\item Alternatively, users may supply their own \code{\link[=vector]{vector}} of custom
box sizes. This vector must be of type "\code{numeric}" and can only contain
positive values. Values which exceed the dimensions of \code{x} are ignored.
}}

\item{periodic}{A Boolean. Determines which boundary algorithm to use, the
classic fixed boundary (default) or the periodic boundary algorithm
introduced by Feagin et al. 2006. The latter is slightly slower but is more
robust to edge effects.}
}
\value{
A \code{\link[=data.frame]{data.frame}} containing box sizes and their
corresponding raw and normalized Λ(r) lacunarity values. Lacunarity is
always computed for box size 1, even if the user supplies a custom
\code{box_sizes} vector that omits it, as this value is required to calculate
normalized lacunarity.
}
\description{
Generates Λ(r) lacunarity curves for a specified set of box sizes, using one
of two versions of the gliding-box algorithm
}
\details{
The raw Λ(r) values depend on the proportion of occupied voxels
within the data space. As a result, it is difficult to compare two spatial
patterns with different occupancy proportions because the curves will begin
at different y-intercepts. This is rectified by normalizing the curve,
typically by log-transforming it and dividing by the lacunarity value at
the smallest box size (i.e. \code{log(Λ(r))/log(Λ(1))}). \code{lacunarity()} outputs
both normalized and non-normalized Λ(r) curves for convenience.
}
\examples{
# generate array
a <- array(data = rep(c(1,0), 125), dim = c(5,5,5))
# calculate lacunarity with default options
lacunarity(a)
# supply custom vector of box sizes
lacunarity(a, box_sizes = c(1,3,5))
# calculate lacunarity at all box sizes using the periodic boundary algorithm
lacunarity(a, box_sizes = "all", periodic = TRUE)
}
