% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacunr-package.R
\docType{package}
\name{lacunr-package}
\alias{lacunr}
\alias{lacunr-package}
\title{lacunr: Fast 3D Lacunarity for LiDAR Voxel Data}
\description{
Calculates 3D lacunarity from voxel data. It is designed to be used with LiDAR point clouds to measure the heterogeneity or "gappiness" of 3-dimensional structures such as forest stands. It provides fast C++ functions to efficiently convert point cloud data to voxels and calculate lacunarity using different variants of Allain & Cloitre's well-known gliding-box algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElliottSmeds/lacunr}
  \item Report bugs at \url{https://github.com/ElliottSmeds/lacunr/issues}
}

}
\author{
\strong{Maintainer}: Elliott Smeds \email{elliott.alfred93@gmail.com} (\href{https://orcid.org/0000-0003-1054-7491}{ORCID})

Other contributors:
\itemize{
  \item J. Antonio Guzmán Q. (Author of original version of voxelize() function) [copyright holder]
}

}
\keyword{internal}
